+++ WallProjections.Test/ViewModels/ViewModelProviderTest.cs
@@ -27,7 +27,7 @@ public class ViewModelProviderTest
         var navigator = new MockNavigator();
         var pythonHandler = new MockPythonHandler();
         using var vmProvider = new ViewModelProvider(navigator, pythonHandler);
-        var displayViewModel = vmProvider.GetDisplayViewModel(new Config(new float[3, 3], new List<Hotspot>()));
+        var displayViewModel = vmProvider.GetDisplayViewModel(new Config(new double[3, 3], new List<Hotspot>()));
         Assert.Multiple(() =>
         {
             Assert.That(displayViewModel, Is.InstanceOf<DisplayViewModel>());
@@ -89,7 +89,7 @@ public class ViewModelProviderTest
     public void GetEditorViewModelTest()
     {
         var hotspot = CreateHotspot(0);
-        var config = new Config(new float[3, 3], new List<Hotspot> { hotspot });
+        var config = new Config(new double[3, 3], new List<Hotspot> { hotspot });
         var navigator = new MockNavigator();
         var pythonHandler = new MockPythonHandler();
         var fileHandler = new MockFileHandler(config);
@@ -282,7 +282,7 @@ public class ViewModelProviderTest
             ImmutableList<string>.Empty,
             ImmutableList<string>.Empty
         );
-        var config = new Config(new float[3, 3], new List<Hotspot> { hotspot });
+        var config = new Config(new double[3, 3], new List<Hotspot> { hotspot });
         var navigator = new MockNavigator();
         var pythonHandler = new MockPythonHandler();
         using var vmProvider = new ViewModelProvider(navigator, pythonHandler);
